formatting and add a test case of std::vector<std::shared_ptr<Vec>> to test vector loading, todo need custom comparator

+++ modules/core/src/MetaObject/core/detail/allocator_policies/Continuous.cpp
@@ -4,7 +4,7 @@ namespace mo
 {
     void ContinuousPolicy::sizeNeeded(int rows, int cols, int elemSize, size_t& size_needed, size_t& stride)
     {
-        stride = static_cast<size_t>(cols * elemSize);
+        stride = static_cast<size_t>(cols) * static_cast<size_t>(elemSize);
         size_needed = stride * static_cast<size_t>(rows);
     }
 }
